Fix bug where game would not load properly.
Closes #9

Don't receive notification that it's your turn after making a new game.

As white, after making a new game, open the PlayView screen. When
another player joins the match and finishes drafting, white will not
receive a notification.

The problem is that when loading in the new game, it would not set the
baseGame to the updated baseGame, because I had mistakenly taken that
out because I thought it was inefficient, since it only needs to be set
after the second ply. I added it back in and added a check to see if the
ply list length is less than two.