Fix issue where matches weren't loaded from PN
This issue was introduced in f21812e, addresses #7

The code now does two checks. First it checks if the deprecated
didFindMatch event was fired, then checks if the
GKTurnBasedMatchmakerViewController is active, and responds accordingly.
This solution handles the now deprecated didFindMatch event on older
devices as well as newer devices that may remove the event all together.